+++ samples/hello/hello_custom_shell.rb
@@ -26,7 +26,8 @@ require 'date'
 # Used to view an email message
 class EmailShell
   # including Glimmer::UI::CustomShell enables declaring as an `email_shell` custom widget Glimmer GUI DSL keyword
-  include Glimmer::UI::CustomShell
+  # Glimmer::UI::CustomShell is a specialized Glimmer::UI::CustomWidget that has `shell` as its `body_root`
+  include Glimmer::UI::CustomShell # aliases: Glimmer::UI::CustomWindow & Glimmer::UI::Application
   
   # multiple options without default values
   options :parent_shell, :date, :subject, :from, :message
@@ -88,26 +89,56 @@ class EmailShell
 end
 
 class HelloCustomShell
-  include Glimmer
+  include Glimmer::UI::Application # aliases: Glimmer::UI::CustomShell & Glimmer::UI::CustomWindow
   
   Email = Struct.new(:date, :subject, :from, :message, keyword_init: true)
-  EmailSystem = Struct.new(:emails, keyword_init: true)
+  EmailSystem = Struct.new(:emails, :selected_email, keyword_init: true)
   
-  def initialize
+  before_body do
     @email_system = EmailSystem.new(
       emails: [
-        Email.new(date: DateTime.new(2029, 10, 22, 11, 3, 0).strftime('%F %I:%M %p'), subject: '3rd Week Report', from: '"Dianne Tux" <dianne.tux@example.com>', message: "Hello,\n\nI was wondering if you'd like to go over the weekly report sometime this afternoon.\n\nDianne"),
-        Email.new(date: DateTime.new(2029, 10, 21, 8, 1, 0).strftime('%F %I:%M %p'), subject: 'Glimmer Upgrade v100.0', from: '"Robert McGabbins" <robert.mcgabbins@example.com>', message: "Team,\n\nWe are upgrading to Glimmer version 100.0.\n\nEveryone pull the latest code!\n\nRegards,\n\nRobert McGabbins"),
-        Email.new(date: DateTime.new(2029, 10, 19, 16, 58, 0).strftime('%F %I:%M %p'), subject: 'Christmas Party', from: '"Lisa Ferreira" <lisa.ferreira@example.com>', message: "Merry Christmas,\n\nAll office Christmas Party arrangements have been set\n\nMake sure to bring a Secret Santa gift\n\nBest regards,\n\nLisa Ferreira"),
-        Email.new(date: DateTime.new(2029, 10, 16, 9, 43, 0).strftime('%F %I:%M %p'), subject: 'Glimmer Upgrade v99.0', from: '"Robert McGabbins" <robert.mcgabbins@example.com>', message: "Team,\n\nWe are upgrading to Glimmer version 99.0.\n\nEveryone pull the latest code!\n\nRegards,\n\nRobert McGabbins"),
-        Email.new(date: DateTime.new(2029, 10, 15, 11, 2, 0).strftime('%F %I:%M %p'), subject: '2nd Week Report', from: '"Dianne Tux" <dianne.tux@example.com>', message: "Hello,\n\nI was wondering if you'd like to go over the weekly report sometime this afternoon.\n\nDianne"),
-        Email.new(date: DateTime.new(2029, 10, 2, 10, 34, 0).strftime('%F %I:%M %p'), subject: 'Glimmer Upgrade v98.0', from: '"Robert McGabbins" <robert.mcgabbins@example.com>', message: "Team,\n\nWe are upgrading to Glimmer version 98.0.\n\nEveryone pull the latest code!\n\nRegards,\n\nRobert McGabbins"),
+        Email.new(
+          date: DateTime.new(2029, 10, 22, 11, 3, 0).strftime('%F %I:%M %p'),
+          subject: '3rd Week Report',
+          from: '"Dianne Tux" <dianne.tux@example.com>',
+          message: "Hello,\n\nI was wondering if you'd like to go over the weekly report sometime this afternoon.\n\nDianne"
+        ),
+        Email.new(
+          date: DateTime.new(2029, 10, 21, 8, 1, 0).strftime('%F %I:%M %p'),
+          subject: 'Glimmer Upgrade v100.0',
+          from: '"Robert McGabbins" <robert.mcgabbins@example.com>',
+          message: "Team,\n\nWe are upgrading to Glimmer version 100.0.\n\nEveryone pull the latest code!\n\nRegards,\n\nRobert McGabbins"
+        ),
+        Email.new(
+          date: DateTime.new(2029, 10, 19, 16, 58, 0).strftime('%F %I:%M %p'),
+          subject: 'Christmas Party',
+          from: '"Lisa Ferreira" <lisa.ferreira@example.com>',
+          message: "Merry Christmas,\n\nAll office Christmas Party arrangements have been set\n\nMake sure to bring a Secret Santa gift\n\nBest regards,\n\nLisa Ferreira"
+        ),
+        Email.new(
+          date: DateTime.new(2029, 10, 16, 9, 43, 0).strftime('%F %I:%M %p'),
+          subject: 'Glimmer Upgrade v99.0',
+          from: '"Robert McGabbins" <robert.mcgabbins@example.com>',
+          message: "Team,\n\nWe are upgrading to Glimmer version 99.0.\n\nEveryone pull the latest code!\n\nRegards,\n\nRobert McGabbins"
+        ),
+        Email.new(
+          date: DateTime.new(2029, 10, 15, 11, 2, 0).strftime('%F %I:%M %p'),
+          subject: '2nd Week Report',
+          from: '"Dianne Tux" <dianne.tux@example.com>',
+          message: "Hello,\n\nI was wondering if you'd like to go over the weekly report sometime this afternoon.\n\nDianne"
+        ),
+        Email.new(
+          date: DateTime.new(2029, 10, 2, 10, 34, 0).strftime('%F %I:%M %p'),
+          subject: 'Glimmer Upgrade v98.0',
+          from: '"Robert McGabbins" <robert.mcgabbins@example.com>',
+          message: "Team,\n\nWe are upgrading to Glimmer version 98.0.\n\nEveryone pull the latest code!\n\nRegards,\n\nRobert McGabbins"
+        ),
       ]
     )
   end
   
-  def launch
-    shell { |shell_proxy|
+  body {
+    shell {
       grid_layout
       
       text 'Hello, Custom Shell!'
@@ -139,22 +170,33 @@ class HelloCustomShell
         }
         
         items <=> [@email_system, :emails, column_properties: [:date, :subject, :from]]
+        selection <=> [@email_system, :selected_email]
         
         on_mouse_up do |event|
-          email = event.table_item.get_data
-          
-          # open a custom email shell
-          email_shell(
-            parent_shell: shell_proxy,
-            date: email.date,
-            subject: email.subject,
-            from: email.from,
-            message: email.message
-          ).open
+          open_email(@email_system.selected_email)
+        end
+        
+        on_key_pressed do |event|
+          if [swt(:cr), swt(:space)].include?(event.keyCode)
+            open_email(@email_system.selected_email)
+          end
         end
       }
-    }.open
+    }
+  }
+  
+  def open_email(email)
+    return if email.nil?
+    
+    # open a custom email shell
+    email_shell(
+      parent_shell: body_root,
+      date: email.date,
+      subject: email.subject,
+      from: email.from,
+      message: email.message
+    ).open
   end
 end
 
-HelloCustomShell.new.launch
+HelloCustomShell.launch